+++ R/coverageCalcOps.R
@@ -1,7 +1,7 @@
 #!/usr/bin/env RScript
 #contributors=c("Gregory Smith", "Nils Jenke", "Michael Gruenstaeudl")
 #email="m_gruenstaeudl@fhsu.edu"
-#version="2024.02.29.2057"
+#version="2024.03.03.0509"
 
 CovCalc <- function(coverageRaw,
                     windowSize = 250,
@@ -34,29 +34,7 @@ CovCalc <- function(coverageRaw,
   return(cov)
 }
 
-getStatsFilePath <- function(sampleName,
-                           plotSpecs) {
-  # Step 1. Check ...
-  if (plotSpecs$isOutput) {
-    outDir <- dirname(plotSpecs$output)
-    tmpDir <- file.path(outDir,
-                        paste(sampleName["sample_name"],
-                              ".tmp",
-                              sep=""))
-  } else {
-    tmpDir <-
-      file.path(".", paste(sampleName["sample_name"],
-                           ".tmp",
-                           sep=""))
-  }
-  # Step 2. Check ...
-  if (dir.exists(tmpDir) == FALSE) {
-    dir.create(tmpDir)
-  }
-  return(tmpDir)
-}
-
-getCovData <- function(regions, genes, analysisSpecs) {
+getCovData <- function(regions, genes) {
   ir_regions <- IRanges::IRanges(
     start = regions$chromStart,
     end = regions$chromEnd,
@@ -96,35 +74,22 @@ getCovData <- function(regions, genes, analysisSpecs) {
     select = "all"
   )
 
-  if (analysisSpecs$isIRCheck) {
-    regions_name <- "Chromosome"
-    regions_start <- "chromStart"
-    regions_end <- "chromEnd"
-  } else {
-    regions_name <- "Source"
-    regions_start <- "srcStart"
-    regions_end <- "srcEnd"
-  }
-
   covData <- list(
     ir_regions = ir_regions,
     ir_genes = ir_genes,
     ir_noncoding = ir_noncoding,
     overlaps_genes = overlaps_genes,
-    overlaps_noncoding = overlaps_noncoding,
-    regions_name = regions_name,
-    regions_start = regions_start,
-    regions_end = regions_end
+    overlaps_noncoding = overlaps_noncoding
   )
   return(covData)
 }
 
-filter_IR_genes <- function(regions, coverageRaw, seqnames, covData) {
+filter_IR_genes <- function(regions, coverageRaw, seqnames, covData, analysisSpecs) {
   ir_genes <- GenomicRanges::GRanges(seqnames = seqnames, covData$ir_genes)
   ir_genes <- GenomicRanges::binnedAverage(ir_genes, coverageRaw, "coverage")
   ir_genes <- as.data.frame(ir_genes)[c("seqnames", "start", "end", "coverage")]
-  regions_name <- covData$regions_name
-  colnames(ir_genes) <- c(regions_name, covData$regions_start, covData$regions_end, "coverage")
+  regions_name <- analysisSpecs$regions_name
+  colnames(ir_genes) <- c(regions_name, analysisSpecs$regions_start, analysisSpecs$regions_end, "coverage")
   ir_genes$coverage <- ceiling(as.numeric(ir_genes$coverage))
   ir_genes[[regions_name]] <- sapply(covData$overlaps_genes, function(x) regions$Band[x])
   ir_genes[[regions_name]] <- gsub("^c\\(\"|\"|\"|\"\\)$", "", as.character(ir_genes[[regions_name]]))
@@ -133,12 +98,12 @@ filter_IR_genes <- function(regions, coverageRaw, seqnames, covData) {
   return(covData)
 }
 
-filter_IR_noncoding <- function(regions, coverageRaw, seqnames, covData) {
+filter_IR_noncoding <- function(regions, coverageRaw, seqnames, covData, analysisSpecs) {
   ir_noncoding <- GenomicRanges::GRanges(seqnames = seqnames, covData$ir_noncoding)
   ir_noncoding <- GenomicRanges::binnedAverage(ir_noncoding, coverageRaw, "coverage")
   ir_noncoding <- as.data.frame(ir_noncoding)[c("seqnames", "start", "end", "coverage")]
-  regions_name <- covData$regions_name
-  colnames(ir_noncoding) <- c(regions_name, covData$regions_start, covData$regions_end, "coverage")
+  regions_name <- analysisSpecs$regions_name
+  colnames(ir_noncoding) <- c(regions_name, analysisSpecs$regions_start, analysisSpecs$regions_end, "coverage")
   ir_noncoding$coverage <- ceiling(as.numeric(ir_noncoding$coverage))
   ir_noncoding[[regions_name]] <- sapply(covData$overlaps_noncoding, function(x) regions$Band[x])
   ir_noncoding[[regions_name]] <- gsub("^c\\(\"|\"|\"|\"\\)$", "", as.character(ir_noncoding[[regions_name]]))
@@ -147,24 +112,24 @@ filter_IR_noncoding <- function(regions, coverageRaw, seqnames, covData) {
   return(covData)
 }
 
-filter_IR_regions <- function(coverageRaw, seqnames, covData) {
+filter_IR_regions <- function(coverageRaw, seqnames, covData, analysisSpecs) {
   ir_regions <- unlist(IRanges::slidingWindows(covData$ir_regions, width = 250L, step = 250L))
   ir_regions <- GenomicRanges::GRanges(seqnames = seqnames, ir_regions)
   ir_regions <- GenomicRanges::binnedAverage(ir_regions, coverageRaw, "coverage")
   chr <- ir_regions@ranges@NAMES
   ir_regions <- as.data.frame(ir_regions, row.names = NULL)[c("seqnames", "start", "end", "coverage")]
   ir_regions["seqnames"] <- chr
-  colnames(ir_regions) <- c(covData$regions_name, covData$regions_start, covData$regions_end, "coverage")
+  colnames(ir_regions) <- c(analysisSpecs$regions_name, analysisSpecs$regions_start, analysisSpecs$regions_end, "coverage")
   ir_regions$coverage <- ceiling(as.numeric(ir_regions$coverage))
 
   covData$ir_regions <- ir_regions
   return(covData)
 }
 
-setLowCoverage <- function(covData) {
+setLowCoverage <- function(covData, analysisSpecs) {
   # ir_regions
   ir_regions <- covData$ir_regions
-  regions_name <- covData$regions_name
+  regions_name <- analysisSpecs$regions_name
   aggFormula <- stats::as.formula(paste("coverage ~", regions_name))
   cov_regions <-
     aggregate(
@@ -199,7 +164,7 @@ setLowCoverage <- function(covData) {
 # adapted from `nilsj9/PlastidSequenceCoverage`
 getCovSummaries <- function(covData,
                             analysisSpecs) {
-  regions_name <- covData$regions_name
+  regions_name <- analysisSpecs$regions_name
 
   covData <- filterCovData(covData,
                            analysisSpecs)
@@ -234,9 +199,9 @@ updateRegionsSummary <- function(covSummaries,
 
 filterCovData <- function(covData,
                           analysisSpecs) {
-  regions_name <- covData$regions_name
-  regions_start <- covData$regions_start
-  regions_end <- covData$regions_end
+  regions_name <- analysisSpecs$regions_name
+  regions_start <- analysisSpecs$regions_start
+  regions_end <- analysisSpecs$regions_end
   windowSize <- analysisSpecs$windowSize
 
   covData$ir_regions <- filterCovDataField(covData$ir_regions,
@@ -340,7 +305,7 @@ calcCovEvenness <- function(df) {
   return(
     df %>%
     dplyr::summarise(
-      evenness = evennessScore(coverage),
+      E_score = evennessScore(coverage),
       .groups = "drop"
     )
   )