+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 CHANGELOG
 ---------
 
-#### Version 1.0.9 (2024.02.29)
+#### Version 1.0.10 (2024.03.03)
+* Dynamic changes to `IRCheck` as `PACVr.complete()` progresses
+  * If no samples from `gbkFile` contain the qualifiers needed for specified analysis, continues with `IRCheck = NA`
+  * If no IR regions are identified, continues with `IRCheck = NA`
+  * If there are missing IR regions needed for synteny test, continues with `IRCheck = 0`
+  * If no synteny is found, continues with `IRCheck = 0`
+
+#### Version 1.0.9 (2024.03.01)
 * Improving variable and file names
 * Correcting output plot specifications
 * Increase logger use, improve logger format